@@ -4,41 +4,42 @@ description: Just Another Workout Timer
 version: 1.11.2+20220317
 
 environment:
-  sdk: ">=2.19.0 <3.0.0"
+  sdk: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_tts: 3.6.3
-  path_provider: 2.0.14
-  flutter_file_dialog: 3.0.0
+  flutter_tts: ^3.7.0
+  path_provider: 2.0.15
+  flutter_file_dialog: ^3.0.1
   pref: 2.7.1
-  fluttertoast: 8.2.1
-  prefs: 4.0.0
-  url_launcher: 6.1.10
+  fluttertoast: ^8.2.2
+  prefs: ^4.1.0+3
+  url_launcher: ^6.1.12
   wakelock: 0.6.2
   soundpool: 2.3.0
-  intl_utils: 2.8.2
-  json_annotation: 4.8.0
+  intl_utils: ^2.8.3
+  json_annotation: 4.8.1
   flutter_phoenix: 1.1.1
-  flutter_native_splash: 2.2.19
-  scrollable_positioned_list: 0.3.5
+  flutter_native_splash: 2.3.0
+  scrollable_positioned_list: 0.3.8
   flutter_localizations:
     sdk: flutter
   uuid: 3.0.7
-  flutter_nord_theme: 2.1.0
-  share_plus: 6.3.1
+#  flutter_nord_theme: 2.1.0
+  share_plus: ^6.3.4
   provider: 6.0.5
+  dynamic_color: ^1.6.6
+  numberpicker: 2.1.2
 
 dev_dependencies:
-  flutter_launcher_icons: 0.12.0
+  flutter_launcher_icons: 0.13.1
   flutter_oss_licenses: 2.0.1
-  json_serializable: 6.6.1
-  build_runner: 2.3.3
-  flutter_lints: 2.0.1
+  json_serializable: ^6.7.1
+  build_runner: ^2.4.6
+  flutter_lints: ^2.0.2
 
-dependency_overrides: 
-  image: ^4.0.10
+#dependency_overrides:
 
 flutter:
   uses-material-design: true
